feat: implement veneer metrics part 1 - implement MetricsApiTracerAdapter (#2631)
port to 1x branch

(cherry picked from commit bf8a7e3e2ae2b7eaa6ad4170f2f3ceb38ddfa48a)
