+++ bigtable-client-core-parent/bigtable-hbase/pom.xml
@@ -59,6 +59,12 @@ limitations under the License.
     </dependency>
 
     <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>bigtable-metrics-api</artifactId>
+      <version>1.21.1-SNAPSHOT</version> <!-- {x-version-update:bigtable-client-core-parent:current} -->
+    </dependency>
+
+    <dependency>
       <groupId>com.google.protobuf</groupId>
       <artifactId>protobuf-java</artifactId>
     </dependency>
