+++ bigtable-client-core-parent/bigtable-metrics-api/src/main/java/com/google/cloud/bigtable/metrics/Counter.java
