+++ bigtable-dataflow-parent/bigtable-beam-import/pom.xml
@@ -149,7 +149,7 @@ limitations under the License.
     <dependency>
       <groupId>commons-codec</groupId>
       <artifactId>commons-codec</artifactId>
-      <version>1.14</version>
+      <version>1.15</version>
     </dependency>
     <dependency>
       <groupId>com.squareup.okhttp</groupId>
