+++ bigtable-dataflow-parent/bigtable-hbase-beam/pom.xml
@@ -27,12 +27,34 @@ limitations under the License.
     and writers in Google Cloud Dataflow.
   </description>
 
+  <dependencyManagement>
+    <dependencies>
+      <!-- fix conflict beam internal between beam-sdks-java-core & beam-model-pipeline -->
+      <dependency>
+        <groupId>com.google.errorprone</groupId>
+        <artifactId>error_prone_annotations</artifactId>
+        <version>2.3.4</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+
   <dependencies>
+    <dependency>
+      <groupId>org.apache.beam</groupId>
+      <artifactId>beam-sdks-java-core</artifactId>
+      <version>${beam.version}</version>
+    </dependency>
     <!-- to use HBaseCoderProviderRegistrar-->
     <dependency>
       <groupId>org.apache.beam</groupId>
       <artifactId>beam-sdks-java-io-hbase</artifactId>
       <version>${beam.version}</version>
+      <!-- HBaseCoderProviderRegistrar discovered during runtime -->
+      <scope>runtime</scope>
+      <!-- Let the beam pipeline choose the appropriate slf4j impl.
+              Since this is the beam universe, we don't have be a drop in replacement
+              for hbase-shaded-client, so we can exclude parts of it that are
+              inappropriate for beam-->
       <exclusions>
         <exclusion>
           <groupId>org.slf4j</groupId>
@@ -41,11 +63,20 @@ limitations under the License.
       </exclusions>
     </dependency>
 
+    <!-- Bigtable Group: ordering doesn't matter since everything is shaded -->
     <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>bigtable-hbase-1.x-shaded</artifactId>
       <version>${project.version}</version>
       <exclusions>
+        <!-- Let the beam pipeline choose the appropriate slf4j impl.
+              Since this is the beam universe, we don't have be a drop in replacement
+              for hbase-shaded-client, so we can exclude parts of it that are
+              inappropriate for beam-->
+        <exclusion>
+          <groupId>org.slf4j</groupId>
+          <artifactId>slf4j-log4j12</artifactId>
+        </exclusion>
         <!-- Workaround MNG-5899 & MSHADE-206. Maven >= 3.3.0 doesn't use the dependency reduced
         pom.xml files when invoking the build from a parent project. So we have to manually exclude
         the dependencies. Note that this works in conjunction with the manually promoted dependencies
@@ -55,23 +86,38 @@ limitations under the License.
           <artifactId>bigtable-hbase-1.x</artifactId>
         </exclusion>
         <exclusion>
+          <groupId>com.google.cloud.bigtable</groupId>
+          <artifactId>bigtable-metrics-api</artifactId>
+        </exclusion>
+        <exclusion>
           <groupId>io.opencensus</groupId>
           <artifactId>*</artifactId>
+        </exclusion>
+        <exclusion>
+          <groupId>io.grpc</groupId>
+          <artifactId>grpc-census</artifactId>
         </exclusion>
       </exclusions>
     </dependency>
 
-    <dependency>
-      <groupId>org.apache.beam</groupId>
-      <artifactId>beam-sdks-java-core</artifactId>
-      <version>${beam.version}</version>
-    </dependency>
+    <!-- HBase Group: contains interfaces bigtable-hbase implement -->
     <dependency>
       <groupId>org.apache.hbase</groupId>
       <artifactId>hbase-shaded-client</artifactId>
       <version>${hbase.version}</version>
+      <exclusions>
+        <!-- Let the beam pipeline choose the appropriate slf4j impl.
+        Since this is the beam universe, we don't have be a drop in replacement
+        for hbase-shaded-client, so we can exclude parts of it that are
+        inappropriate for beam-->
+        <exclusion>
+          <groupId>org.slf4j</groupId>
+          <artifactId>slf4j-log4j12</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
 
+    <!-- Misc Group -->
     <dependency>
       <groupId>com.google.code.findbugs</groupId>
       <artifactId>jsr305</artifactId>
@@ -91,6 +137,7 @@ limitations under the License.
       <version>1.20</version>
     </dependency>
 
+    <!-- Test Group -->
     <dependency>
       <groupId>org.mockito</groupId>
       <artifactId>mockito-core</artifactId>
