+++ bigtable-hbase-1.x-parent/bigtable-hbase-1.x-hadoop/pom.xml
@@ -68,6 +68,11 @@ limitations under the License.
       <version>${jsr305.version}</version>
     </dependency>
     <dependency>
+      <groupId>com.google.cloud.bigtable</groupId>
+      <artifactId>bigtable-metrics-api</artifactId>
+      <version>${project.version}</version> <!-- {x-version-update:bigtable-client-core-parent:current} -->
+    </dependency>
+    <dependency>
       <groupId>commons-logging</groupId>
       <artifactId>commons-logging</artifactId>
       <version>${commons-logging.version}</version>
@@ -75,7 +80,7 @@ limitations under the License.
     <dependency>
       <groupId>io.dropwizard.metrics</groupId>
       <artifactId>metrics-core</artifactId>
-      <version>${dropwizard.metrics.version}</version>
+      <version>${hbase1-metrics.version}</version>
     </dependency>
 
     <!-- this is used only for test, but we need this as compile scope for repackaging-->
