+++ bigtable-hbase-1.x-parent/bigtable-hbase-1.x-tools/pom.xml
@@ -47,7 +47,7 @@
     <dependency>
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
-      <version>30.0-jre</version>
+      <version>30.1.1-jre</version>
     </dependency>
     <dependency>
       <groupId>org.slf4j</groupId>
