+++ pom.xml
@@ -56,8 +56,9 @@ limitations under the License.
     <compileSource.1.8>1.8</compileSource.1.8>
 
     <!-- core dependency versions -->
-    <bigtable.version>1.19.2</bigtable.version>
-    <dropwizard.metrics.version>3.2.6</dropwizard.metrics.version>
+    <bigtable.version>1.26.1</bigtable.version>
+    <hbase1-metrics.version>3.1.2</hbase1-metrics.version>
+    <hbase2-metrics.version>3.2.6</hbase2-metrics.version>
     <!--  keeping at this version to align with hbase-->
     <slf4j.version>1.7.25</slf4j.version>
     <commons-logging.version>1.2</commons-logging.version>
@@ -78,7 +79,7 @@ limitations under the License.
     <!-- TODO: check if commons-codec was transitively updated to 1.13 and okhttp was updated to 2.7.5 when upgrading-->
     <beam.version>2.24.0</beam.version>
     <!-- referred from bigtable-beam-import and bigtable-emulator -->
-    <guava.version>30.0-android</guava.version>
+    <guava.version>30.1.1-android</guava.version>
     <beam-guava.version>20.0</beam-guava.version>
     <beam-auto-value.version>1.7</beam-auto-value.version>
     <gcs-guava.version>29.0-jre</gcs-guava.version>
